数据类型对齐

+++ src/yolo11_pkg/scripts/onnx.py
@@ -64,7 +64,7 @@ class YOLOv8:
         """Preprocess the input image."""
         img = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
         img = cv2.resize(img, (self.input_width, self.input_height))
-        img = img.astype(np.float32) / 255.0
+        img = img.astype(np.float16) / 255.0
         img = np.transpose(img, (2, 0, 1))  # Channel first
         return np.expand_dims(img, axis=0)
